sistema-manejo-pedidos/Procesamiento/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-login
aiokafka
requests
xxhash